Refactor: Utils 이름 변경
